requirements.txt
# Django and REST Framework
Django==5.2.7
djangorestframework==3.15.2
django-cors-headers==4.6.0

# Blockchain Integration
web3==7.7.0
eth-tester==0.13.0b1
py-solc-x==2.0.3

# Redis for caching and rate limiting
redis==5.2.1
django-redis==5.4.0

# Property-Based Testing
hypothesis==6.122.4
pytest==8.3.4
pytest-django==4.9.0
pytest-cov==6.0.0

# Additional Testing Tools
fakeredis==2.26.2

# Utilities
python-dotenv==1.0.1
